clk: ti: am437x: fix parent of synctimer 32k
The external 32KHz RTC clock source may not always be available
on board like in the case of ePOS EVM. By default sync timer, which
is used as clock source, feeds of this clock. This is a problem.
Change the parent of sync timer to PER PLL 32KHz clock instead
which is always present. This has a side effect that in low power
modes, sync timer will stop.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>