omapdss: DRA7x DPI: set dpi's pll to NONE when disconnecting
We assign a dpll to a DPI interface when we connect it to an overlay manager.
However, we unassign the dpll when we disable DPI. This causes the DPI driver
to not use the dpll the next time it's enabled.

Unassigning the dpll in disconnect() rather than disable() prevents this.

Signed-off-by: Archit Taneja <archit@ti.com>