input: touchscreen: fix spelling mistake in TSC/ADC DT binding
There was a spelling mistake on TSC/ADC binding where
"coordinate" was spelled as "coordiante".

We can't simply fix the error due to DT being an ABI,
the approach taken was to first use correct spelling
and if that fails, fallback to miss-spelled version.

It's unfortunate that has creeped into the tree.

Signed-off-by: Felipe Balbi <balbi@ti.com>