ARM: dts: am33xx: Add clkdiv32k_ick to clk_24mhz_clkdm clockdomain
The debounce clock for GPIO1 is a part of clk_24mhz_clkdm clockdomain.
Not maintaining this association in DT will turn the clockdomain inactive
thus causing the debounce clock to GPIO1 to turn off. This results in crash
during boot when gpio-keys driver is used.

commit ce5b27412d45ca1f26899578f7e6986b1676e9c5 (ARM: dts: am33xx clock
data) introduced nodes, but clkdiv32k_ick is not associated with a clock
domain.

We change gate-clock to ti,gate-clock for the clkdiv32k_ick as
suggested by Tero and make the association to the clk_24mhz_clkdm.

Signed-off-by: Joel Fernandes <joelf@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Rajendra Nayak <rnayak@ti.com>
Acked-by: Santosh Shilimkar <santosh.shilimkar@ti.com>