mtd: tests: oobtest: fix test failures for oobsize which are not multiples of 4
If oobsize is not multiple of 4 then the generated data to write to flash
and data generated to verify are not same. This imapcts test cases 2,3 and 5.
If mtd->ecclayout->oobavail is not a multiple of 4 then while generating psuedo
random numbers to write to flash prandom_bytes_state () generates the extra
random bytes which wont be done if we do for the whole length. So check
for mtd->ecclayout->oobavail if its multiple of 4 then generate for full len
(mtd->ecclayout->oobavail * pgcnt)
else generate prandom of mtd->ecclayout->oobavail bytes pgcnt times.

Signed-off-by: George Cherian <george.cherian@ti.com>