OMAPDSS: HDMI5: mask & clear irqs properly
The hdmi_core_mask_interrupts() function is supposed to mask all the
various interrupts from the HDMI core. This was not done correctly, and
if the CORE_IRQ was enabled, it would result in an interrupt flood.

Fix the function to first mask all the interrupts and then clear the
interrupt status bits.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>