OMAPDSS: HDMI5: workaround for hdmi rxdet
When we set the PHY to LDOON and we don't have a HDMI cable connected,
we always see a LINK_CONNECT and LINK_DISCONNECT interrupt. This is a
problem because our irq handler will set the PHY power to OFF and then
to LDOON when getting both irqs at the same time, leading to infinite
amount of interrupts.

As a workaround, we can override the RXDET line for the duration of the
PHY power handling, so that those bogus interrupts will be hidden.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>