OMAPDSS: HDMI5: improve EDID reading
The EDID reading in the driver is rather bad. It does not check any
error status, and even worse, it doesn't even check if any data was
received. The driver just waits for a specified time, and presumes a
byte was received after the wait.

Rewrite the EDID reading to actually look at the DONE and ERROR status
bits when reading EDID.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>