OMAPDSS: HDMI5: calculate I2CM timings more properly
The HDMI I2CM timings are currently hardcoded in the driver, by just
having magic numbers that are programmed to the registers.

This patchs calculates them more properly. We still use hardcoded start
values for DSS L3 ICLK rate and SFR_DIV, as we don't have a way to get
the DSS L3 ICLK rate at the moment.

This should be fixed when the DSS L3 ICLK rate is available.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>