OMAPDSS: HDMI5: use no-idle when reading edid
The HDMI core uses DSS L3 ICLK as a functional clock (or that is what it
seems like). This means that if the ICLK may be gated automatically
resulting in very slow EDID reading.

Set the HDMI's IDLEMODE to no-idle for the duration of the EDID reading
so that the ICLK will not be gated.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Cc: Tero Kristo <t-kristo@ti.com>
Cc: Nishanth Menon <nm@ti.com>