ARM: dts: OMAP5: Pass the timer clock source info from DT
The dmtimer platform driver/api expects the timer source clock to have
an alias of 'timer_sys_ck'. The current alias tables as part of the clock
driver do not work with DT and hence the dmtimer api to set the timer source
to 'timer_sys_ck' fail.

Fix this by passing the info about the timer clock source from DT and get rid
of all the alias entries in the table currently part of the clock driver.

Signed-off-by: Rajendra Nayak <rnayak@ti.com>
Suggested-by: Nishanth Menon <nm@ti.com>