ARM: head.S: avoid ldm access crossing 1K page boundary
Due to a bug in an early internal version of AM43x engineering sample,
there is an issue with ldm access crossing 1K page in uncached memory.
This patch avoids it with a very negligible performance penalty.

The engineering sample having this issue is TI-internal, this patch is
used to enable internal development. The production version of AM43x
does not have this bug and this patch can be reverted at that time.

Signed-off-by: Afzal Mohammed <afzal@ti.com>