clk: ti: clk-7xx: Correct ABE DPLL configuration
ABE DPLL frequency need to be lowered from 361267200
to 180633600 to facilitate the ATL requironments.
The dpll_abe_m2x2_ck clock need to be set to double
of ABE DPLL rate in order to have correct clocks
for audio.

clk: ti: clk-7xx: Configure abe_m2x2 clock speed

It need to be set to x2 of ABE dpll

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>