v4l: ti-vpe: Create a scaler library
Create the scaler block related configurations and register definitions as a
library. This is done as the same scaler block is used in both VIP and VPE.

The vpe_dev device holds a sc_data handle, which it can reference to access
the registers, or call scaler related helper functions.

Signed-off-by: Archit Taneja <archit@ti.com>