ARM: dts: AM33XX: Add efuse_sma entry for OPP and add higher OPPs
Add an entry for efuse_sma which configures OPPs on 33xx. Within this
nodes are defined with opp-modifier propety that are defined as a list
of frequency, revision, offset from base register, and efuse value.
The CPU node (or any node) passes a phandle to the appropriate child
node of the efuse node.

This patch also adds the ES2.x OPPs for 335x.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>