boot: dts: Add wlan support for Aegis-epos
The wilink module is a combo wireless connectivity sdio card based on
Texas Instrument's wl18xx solution.
It is a 4-wire, 1.8V, embedded sdio wlan device with an external irq
line and is power-controlled by a gpio-based fixed regulator.

Add support for the wlan capabilities of this module by muxing the
relevant mmc and gpio lines, and setting the required platform data.

Signed-off-by: Vijay Badawadagi <bvijay@ti.com>