ARM: OMAP4: hwmod data: add back irq data for mmus
The OMAP IOMMU driver is not yet converted for device-tree, and
as such still needs the irq information from the corresponding
platform device for the driver probe to be successful. This data
was cleaned up previously as part of the patch,
3b9b101 "ARM: OMAP4: hwmod data: Clean up the data file".

Add this irq data back for the IPU and DSP MMUs to make the
iommu driver functional again.

Signed-off-by: Suman Anna <s-anna@ti.com>