CLK: TI: OMAP5: initialize IVA DPLL and associated output clocks
The IVA DPLL may not be usually configured by bootloader (e.g. u-boot
configures it only if CONFIG_SYS_CLOCKS_ENABLE_ALL is enabled), and
this results in various issues when trying to use a DSP or IVAHD,
whose root clocks are sourced from this DPLL. Initialize the IVA DPLL
at boot time to avoid the functional issues caused when using either
of these devices if the DPLL is not-locked. The DPLL will automatically
transition into a low-power stop mode when the associated output clocks
are not utilized or gated automatically.

The reset values of the dividers H11 & H12 are identical to each other,
but are different at each OPP. The reset values also may not match a
specific OPP. So, the derived output clocks from the IVA DPLL have to
be initialized as well (OPP100 values chosen currently) to avoid
initializing these divider outputs to incorrect frequencies.

Signed-off-by: Suman Anna <s-anna@ti.com>