ARM: dra7: Add delay values for Pin control
To Guarantee IO timings on certain pins of dra7, we need to add a
specific delay to each corresponding control register. This delay
configuration support for vin1a_fld0 port is added. Delay can be
configured at the time pinmux is configured for vin1a_fld0.

DSS_VOUT3 muxed from this pin is not used currently. When we need
to use this pin for DSS_VOUT3 clock, this additional configuration
is is to be written to the pin's control register.

Signed-off-by: Sathya Prakash M R <sathyap@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>