ARM: dts: am43x-epos-evm: reduce i2c0 busspeed to 100KHz
Epos board has an option of plugging in an camera board. However, the
camera board in question uses txs0102[1] level shifter. This level
shifter has an strong internal pull up of 10k - as a result of which
the i2c bus characteristics drastically changes. More details in [2].

Due to the change in bus characteristics, TPS65218 may miscorelate
voltage change or confirmation requests to the wrong register and fail
to change voltage. This has the following side effect:

When transitioning from 300MHz to 720MHz or viceversa, it is possible
to end up missing the critical sequence of ensuring voltage level
is mandatory to be at the level of relevant OPP. Due to the bus
characteristics change, this over a multiple iteration, result in
VDD_MPU at OPP50 voltage and frequency at 720MHz - resulting in a hang
of the system.

Reducing the I2C speed down to 100KHz seems to ensure every peripheral
continues to be within spec ranges with and without the camera board
connected. The other alternative will be to create a custom dts for
epos-camera board configuration, but that will result in avoidable
confusion.

NOTE: this could have impact on performance on the other peripherals
on i2c0 such as touchscreen etc. Better optimization might be possible
with adequate board modification and detailed bus characterization.

[1] http://www.ti.com/lit/ds/symlink/txs0102.pdf
[2] http://e2e.ti.com/support/interface/etc_interface/f/390/t/149757.aspx

Reported-by: Yan Liu <yan-liu@ti.com>
Reported-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Acked-by: Felipe Balbi <balbi@ti.com>