v4l: ti-vpe: load scaler coefficients
The scaler IP requires to be loaded with coefficients from memory through
VPDMA.

The horizontal and vertical scaler each require a set of scaler coefficients.
The horizontal polyphase scaler requires luma and chroma coefficients for a 32
phase and 8 tap filter. Similarly, the vertical scaler requires coefficients for
a 5 tap filter.

In order to load the scaler coefficients via VPDMA, a configuration descriptor
is used in block mode. The payload for the descriptor is the scaler coefficients
copied to memory. The coefficients need to be placed in memory in a particular
order, this is taken care by the functions sc_set_hs and sc_set_vs_coefficients.

The choice of the scaler coefficients depends on the downscale ratio. In the
case of horizontal downscaling, we need to consider the change in ratio cause by
decimation. In the case of vertical scaling, the VPE driver should provide the
source height based on whether the de-interlacer is used or not.

Signed-off-by: Archit Taneja <archit@ti.com>