clk: ti: am437x: enable clkout2 by default
Add clk node for clkout2 and enable it by
default. clkout2 provides clock to onboard
peripherals like WLAN. This clock needs to
be enabled for WLAN to be probed. Enabling
clkout2 by default and leaving it enabled
does not affect entry into DeepSleep state
since it comes from power domain which is
expected to remain on during low power modes
(RTC domain and the PRCM itself).

This is similar to what is done on AM335x.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>