v4l: ti-vpe: enable basic scaler support
Add the required scaler register configurations which lets us perform linear
scaling for the supported range of horizontal and vertical scaling ratios.

The horizontal scaler performs polyphase scaling using it's 8 tap 32 phase
filter, decimation is performed when downscaling passes beyond 2x or 4x.

The vertical scaler performs polyphase scaling using it's 5 tap 32 phase
filter, it switches to a simpler form of scaling using the running average
filter when the downscale ratio is more than 4x.

Many of the scaler features like peaking, trimming and non-linear scaling aren't
implemented for now. Only those register fields are configured which are
required for basic scaling operation.

The function to configure scaler registers takes the sc_data handle, the source
and destination directions, and the scaler address data block offsets for the
current context so that they can be configured.

Signed-off-by: Archit Taneja <archit@ti.com>