ARM: OMAP: remoteproc: adjust the CMA pools for DRA7
The CMA base addresses and sizes are adjusted to meet the current
usecase requirements for the various processors for DRA7XX. The
IPU2 size is decreased by 112MB to account for the removal of
the OMAP camera stack, and the sizes for DSP1 and IPU1 are increased
to meet the current image needs. The CMA base addresses are also
adjusted and calculated using 0x95800000 as the starting base for
IPU2 and using the higher addresses for subsequent processors based
on the changed CMA pool sizes. The current CMA pool would end at
0x9F800000.

NOTE:
1. This requires the BIOS-side images to use matching addresses
   in the resource table, so older images with different resource
   tables may not boot after this change.
2. This change also helps with the enablement of fdt_high relocation
   in ti-u-boot, which causes the DT blob to occupy memory in the last
   MB before 0xA0000000 (fdt_high value) and thereby resulting in the
   failure of one of the rproc CMA pools.

Signed-off-by: Suman Anna <s-anna@ti.com>