ARM: dts: add AM33XX vdd core opp50 suspend for am335x-evm
This adds a sleep and wake sequence to set the VDD core voltage to the
OPP50 level, 0.950V. This saves power during suspend. The sequences are
specific to the am335x-evm layout and PMIC, the TPS65910. The sequences
are written out by the Cortex-M3.

Tested on am335x-evm board for volatge changes back and forth during
suspend and resume.

Signed-off-by: Russ Dill <Russ.Dill@ti.com>
Signed-off-by: Keerthy <j-keerthy@ti.com>