omapdss: dss dplls: Ignore PLL_PWR_STATUS on DRA7x
There is a bug in DRA7x hardware which prevents the correct update of
PLL_PWR_STATUS when the power state of DPLL_VIDEO1/2 is changed.

Currently, the driver reports the error but still proceeds ahead. Modify the
code such that we don't read the STATUS field at all and add a small delay
for the power state to change. The dss dpll driver is currently used only for
DRA7x, the fix can be added in a more cleaner manner once it is used for DSI
PLLs on previous OMAP revisions too.

Signed-off-by: Archit Taneja <archit@ti.com>