ARM: DRA7: hwmod data: remove main_clk for mailboxes
There is no separate functional clock that needs to be enabled
for the mailbox module instances in DRA7xx, like in other OMAP
SoCs. The L4CFG_L3_GICLK interface clock serves as both the
functional and interface clock to the modules, and this clock
is defined through the associated omap_hwmod_ocp_if definitions,
so remove the main_clk field from the hwmod definitions.

This is inline with the hwmod definitions on OMAP4 and OMAP5,
and also fixes the warning traces with the DT clock series from
the platform base tree. The traces look like
"platform xxxxxxxx.mailbox: FIXME: clock-name 'fck' DOES NOT exist in dt!"

Signed-off-by: Suman Anna <s-anna@ti.com>