ARM: DRA7: hwmod data: remove main_clk for spinlock
There is no separate functional clock that needs to be enabled
for the hardware spinlock module in DRA7xx, like in other OMAP
SoCs. The L4CFG_L3_GICLK interface clock serves as both the
functional and interface clock to the module, and is defined
through the associated omap_hwmod_ocp_if definition, so remove
the main_clk field from the hwmod definition.

This is inline with the hwmod definition on OMAP4 and OMAP5,
and also fixes the warning trace:
"platform 4a0f6000.spinlock: FIXME: clock-name 'fck' DOES NOT exist in dt!"

Signed-off-by: Suman Anna <s-anna@ti.com>