ARM: AMx3x: hwmod data: remove main_clk for spinlock
There is no separate functional clock that needs to be enabled
for the hardware spinlock module in AM33xx & AM43xx, like in other
OMAP SoCs. The L4PER_CLK interface clock serves as both the
functional and interface clock to the module, and is defined
through the associated omap_hwmod_ocp_if definition, so remove
the main_clk field from the hwmod definition.

This is inline with the hwmod definition on newer OMAP SoCs,
and also fixes the warning trace:
"platform 480ca000.spinlock: FIXME: clock-name 'fck' DOES NOT exist in dt!"

Signed-off-by: Suman Anna <s-anna@ti.com>