iio: ti_am335x_adc: Poll fifo1Thresh interrupt instead of ADCStatus for idle
The ADCSTATUS register can not be polled while waiting for idle while
the TSC is used because it might happen that the sequencer never gets
idle leading to -EBUSY.
Instead we setup the FIFO threshold and wait for it to fill. The wait
period takes about 500us. If we reduce the number of slots to process to
only one because we only interrested into one entry then the wait
perioud is reduced to approx. 125us.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Jeff Lance <j-lance1@ti.com>