arm: OMAP: change audio drivers from modules to builtin in omap2plus_defconfig
Change the default setting of various audio related tristate config options
from modules to builtins.  This is to work around an issue where the audio
deferred probes appear to not be triggering correctly when certain boards
are booted with tftp/nfs.

v2:  put OMAP3 Pandora back to a module to debloat the kernel a tiny
amount.

Signed-off-by: Darren Etheridge <detheridge@ti.com>