usb: musb: am335x: Make musb_am335x as built-in type
When the PHY device is deleted during rmmod musb_am335x,
all its managed resources are deleted in __device_release_driver()
after the driver's .remove() hook is called.

This means the usb_phy device structure will no longer be valid and
the usb_phy->ops cannot be called.

We avoid this situation by making the musb_am335x driver as
built-in, so that the PHY device can never be deleted.

A more proper fix would be to use a more reliable PHY framework
that takes care of such cases.

Fixes: D-01175: AM335x - rmmod musb_am335x causes crash

Signed-off-by: Roger Quadros <rogerq@ti.com>