Revert "usb: musb: musb_core: Fix error path in probe"
This reverts commit 9f9bd9296c7f6a116e774fec4b0964a17828670c
which was trying to address an issue introduced by
commit 90c17b09d "usb: musb: musb_core: Destroy dma controller only once".
The issue was that removing and reloading musb_dsps module caused a
backtrace. This was because it was placing the dma_controller_destroy()
in musb_free() instead of musb_remove().

Now that commit de98f7a5 "TI-Integration: usb: musb: fix miss merge in
musb_core" is in tree and hopefully commit 90c17b09 will be reverted
then we no longer need the patch which is being reverted and we're
aligned with mainline.

Signed-off-by: Roger Quadros <rogerq@ti.com>