bus: omap_l3_noc: fix unclearable error handling
Errors that cannot be cleared (determined by reading status register)
are currently handled by masking it. Status register is understood to be
raw status rather than result of masking, based on recent issue reported
by Vaibhav Hiremath (I could not make out from the documentation that it
is raw status).

As that is the case, if after the occurance of an unclearable error,
another clearable L3 error happens that has lower binary weightage in
flagmux register (compared to the unclearable error that has occured),
and as status register is raw, isr will think that interrupt is raised
due to unclearable error(even though it was masked). This causes the new
error that could have been cleared, uncleared, resulting in isr getting
called repeatedly.

Fix it by masking raw status with masked ones before handling the error.

Reported-by: Vaibhav Hiremath <hvaibhav@ti.com>
Signed-off-by: Afzal Mohammed <afzal@ti.com>
Tested-by: Vaibhav Hiremath <hvaibhav@gmail.com>