ARM: dts: AM4372: Add efuse device entry for OPP and add higher OPPs
Add an entry for device_attribute which configures OPPs on 43xx. Within
this, nodes are defined with the opp-modifier propety that are defined as
a list of frequency, revision, offset from base register, and efuse value.
The CPU node passes a phandle to the appropriate child node to get the
correct table.

This patch also adds higher eFused OPPs for AM4372.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>